型エラー修正・新規島登録画面の活動内容コンポーネント化完成
