+++ src/components/createIsland/detail/handleIslandDetailChange.tsx
@@ -1,0 +1,13 @@
+// 一文字でも入力されたらエラー解消
+
+export default function HandleDetailChange({ setDetail, error, setError }) {
+  const handleIslandDetailChange = (e) => {
+    setDetail(e.target.value);
+    // 一文字でも入力されたらエラー削除
+    if (error) {
+      setError("");
+    }
+  };
+
+  return handleIslandDetailChange;
+}
