+++ src/components/createIsland/islandName/handleIslandNameChange.tsx
@@ -1,7 +1,5 @@
 // 一文字でもinputタグに入力されたらエラーを解消
 
-import { Dispatch, SetStateAction } from "react";
-
 export default function HandleIslandNameChange({
   islandNameError,
   setIslandName,
